+++ Example/PaymentSheet Example/PaymentSheetUITest/CustomerSheetUITest.swift
@@ -17,7 +17,9 @@ class CustomerSheetUITest: XCTestCase {
         continueAfterFailure = false
 
         app = XCUIApplication()
-        app.launchEnvironment = ["UITesting": "true"]
+        app.launchEnvironment = ["UITesting": "true",
+                                 "USE_PRODUCTION_FINANCIAL_CONNECTIONS_SDK": "true",
+        ]
         app.launch()
     }
 
@@ -177,6 +179,47 @@ class CustomerSheetUITest: XCTestCase {
         let selectButtonFinal = app.staticTexts["None"]
         XCTAssertTrue(selectButtonFinal.waitForExistence(timeout: 60.0))
     }
+    func testAddTwoPaymentMethod_setupIntent() throws {
+        var settings = CustomerSheetTestPlaygroundSettings.defaultValues()
+        settings.customerMode = .new
+        settings.applePay = .on
+        settings.paymentMethodMode = .setupIntent
+        loadPlayground(
+            app,
+            settings
+        )
+        let selectButton = app.staticTexts["None"]
+        XCTAssertTrue(selectButton.waitForExistence(timeout: 60.0))
+        selectButton.tap()
+
+        app.staticTexts["+ Add"].tap()
+
+        try! fillCardData(app, postalEnabled: true)
+        app.buttons["Save"].tap()
+
+        let cardPresence = app.staticTexts["••••4242"]
+        XCTAssertTrue(cardPresence.waitForExistence(timeout: 60.0))
+
+        app.staticTexts["+ Add"].tap()
+
+        try! fillCardData(app, cardNumber: "5555555555554444", postalEnabled: true)
+        app.buttons["Save"].tap()
+
+        let cardPresence4444 = app.staticTexts["••••4444"]
+        XCTAssertTrue(cardPresence4444.waitForExistence(timeout: 60.0))
+
+        let cardPresence4242 = app.staticTexts["••••4242"]
+        XCTAssertTrue(cardPresence4242.waitForExistence(timeout: 60.0))
+
+        let closeButton = app.buttons["Close"]
+        XCTAssertTrue(closeButton.waitForExistence(timeout: 60.0))
+        closeButton.tap()
+
+        dismissAlertView(alertBody: "Success: payment method not set, canceled", alertTitle: "Complete", buttonToTap: "OK")
+
+        let selectButtonFinal = app.staticTexts["None"]
+        XCTAssertTrue(selectButtonFinal.waitForExistence(timeout: 60.0))
+    }
     func testAddPaymentMethod_createAndAttach_reInitAdddViewController() throws {
         var settings = CustomerSheetTestPlaygroundSettings.defaultValues()
         settings.customerMode = .new
@@ -209,6 +252,47 @@ class CustomerSheetUITest: XCTestCase {
         let backButton = app.buttons["Back"]
         XCTAssertTrue(backButton.waitForExistence(timeout: 60.0))
         backButton.tap()
+
+        let closeButton = app.buttons["Close"]
+        XCTAssertTrue(closeButton.waitForExistence(timeout: 60.0))
+        closeButton.tap()
+
+        dismissAlertView(alertBody: "Success: payment method not set, canceled", alertTitle: "Complete", buttonToTap: "OK")
+
+        let selectButtonFinal = app.staticTexts["None"]
+        XCTAssertTrue(selectButtonFinal.waitForExistence(timeout: 60.0))
+    }
+    func testAddTwoPaymentMethod_createAndAttach() throws {
+        var settings = CustomerSheetTestPlaygroundSettings.defaultValues()
+        settings.customerMode = .new
+        settings.applePay = .on
+        settings.paymentMethodMode = .createAndAttach
+        loadPlayground(
+            app,
+            settings
+        )
+        let selectButton = app.staticTexts["None"]
+        XCTAssertTrue(selectButton.waitForExistence(timeout: 60.0))
+        selectButton.tap()
+
+        app.staticTexts["+ Add"].tap()
+
+        try! fillCardData(app, postalEnabled: true)
+        app.buttons["Save"].tap()
+
+        let cardPresence = app.staticTexts["••••4242"]
+        XCTAssertTrue(cardPresence.waitForExistence(timeout: 60.0))
+
+        app.staticTexts["+ Add"].tap()
+
+        try! fillCardData(app, cardNumber: "5555555555554444", postalEnabled: true)
+        app.buttons["Save"].tap()
+
+        let cardPresence4444 = app.staticTexts["••••4444"]
+        XCTAssertTrue(cardPresence4444.waitForExistence(timeout: 60.0))
+
+        let cardPresence4242 = app.staticTexts["••••4242"]
+        XCTAssertTrue(cardPresence4242.waitForExistence(timeout: 60.0))
 
         let closeButton = app.buttons["Close"]
         XCTAssertTrue(closeButton.waitForExistence(timeout: 60.0))
@@ -333,7 +417,9 @@ class CustomerSheetUITest: XCTestCase {
         XCTAssertTrue(selectButton.waitForExistence(timeout: 60.0))
         selectButton.tap()
 
-        app.staticTexts["+ Add"].tap()
+        let addButton = app.staticTexts["+ Add"]
+        XCTAssertTrue(addButton.waitForExistence(timeout: 60.0))
+        addButton.tap()
 
         try! fillCardData(app, postalEnabled: true)
         app.buttons["Save"].tap()
@@ -404,7 +490,54 @@ class CustomerSheetUITest: XCTestCase {
 
         dismissAlertView(alertBody: "Success: ••••4242, canceled", alertTitle: "Complete", buttonToTap: "OK")
     }
+    // TODO: Uncomment when enabling USBankAccount
+/*
+    func testCustomerSheetStandard_applePayOff_addUSBankAccount() throws {
+        var settings = CustomerSheetTestPlaygroundSettings.defaultValues()
+        settings.customerMode = .new
+        settings.applePay = .off
+        settings.paymentMethodTypes = .card_usbank
+        loadPlayground(
+            app,
+            settings
+        )
 
+        let selectButton = app.staticTexts["None"]
+        XCTAssertTrue(selectButton.waitForExistence(timeout: 60.0))
+        selectButton.tap()
+
+        let usBankAccountPMSelectorButton = app.staticTexts["US Bank Account"]
+        XCTAssertTrue(usBankAccountPMSelectorButton.waitForExistence(timeout: 60.0))
+        usBankAccountPMSelectorButton.tap()
+
+        try! fillUSBankData(app)
+
+        let continueButton = app.buttons["Continue"]
+        XCTAssertTrue(continueButton.waitForExistence(timeout: 60.0))
+        continueButton.tap()
+
+        // Go through connections flow
+        app.buttons["Agree and continue"].tap()
+        app.staticTexts["Test Institution"].forceTapElement()
+        app.staticTexts["Success"].waitForExistenceAndTap(timeout: 10)
+        app.buttons["Link account"].tap()
+        XCTAssertTrue(app.staticTexts["Success"].waitForExistence(timeout: 10))
+        app.buttons.matching(identifier: "Done").allElementsBoundByIndex.last?.tap()
+
+        let testBankLinkedBankAccount = app.staticTexts["StripeBank"]
+        XCTAssertTrue(testBankLinkedBankAccount.waitForExistence(timeout: 60.0))
+
+        let saveButton = app.buttons["Save"]
+        XCTAssertTrue(saveButton.waitForExistence(timeout: 60.0))
+        saveButton.tap()
+
+        let confirmButton = app.buttons["Confirm"]
+        XCTAssertTrue(confirmButton.waitForExistence(timeout: 60.0))
+        confirmButton.tap()
+
+        dismissAlertView(alertBody: "Success: ••••6789, selected", alertTitle: "Complete", buttonToTap: "OK")
+    }
+*/
     func presentCSAndAddCardFrom(buttonLabel: String, tapAdd: Bool = true) {
         let selectButton = app.staticTexts[buttonLabel]
         XCTAssertTrue(selectButton.waitForExistence(timeout: 60.0))