+++ Stripe/StripeiOSTests/PaymentSheet+APITest.swift
@@ -13,6 +13,7 @@ import XCTest
 @testable@_spi(STP) import StripeCore
 @testable@_spi(STP) import StripePayments
 @testable@_spi(STP) import StripePaymentSheet
+@testable@_spi(STP) import StripeUICore
 
 class PaymentSheetAPITest: XCTestCase {