+++ StripeIdentity/StripeIdentity/Source/API Bindings/Models/VerificationPageDataUpdate/VerificationPageCollectedData.swift
@@ -21,6 +21,8 @@ extension StripeAPI {
         private(set) var dob: VerificationPageDataDob?
         private(set) var name: VerificationPageDataName?
         private(set) var address: RequiredInternationalAddress?
+        private(set) var phone: VerificationPageDataPhone?
+        private(set) var phoneOtp: String?
 
         init(
             biometricConsent: Bool? = nil,
@@ -31,7 +33,9 @@ extension StripeAPI {
             idNumber: VerificationPageDataIdNumber? = nil,
             dob: VerificationPageDataDob? = nil,
             name: VerificationPageDataName? = nil,
-            address: RequiredInternationalAddress? = nil
+            address: RequiredInternationalAddress? = nil,
+            phone: VerificationPageDataPhone? = nil,
+            phoneOtp: String? = nil
         ) {
             self.biometricConsent = biometricConsent
             self.face = face
@@ -42,6 +46,8 @@ extension StripeAPI {
             self.dob = dob
             self.name = name
             self.address = address
+            self.phone = phone
+            self.phoneOtp = phoneOtp
         }
     }
 }
@@ -62,7 +68,9 @@ extension StripeAPI.VerificationPageCollectedData {
             idNumber: otherData.idNumber ?? self.idNumber,
             dob: otherData.dob ?? self.dob,
             name: otherData.name ?? self.name,
-            address: otherData.address ?? self.address
+            address: otherData.address ?? self.address,
+            phone: otherData.phone ?? self.phone,
+            phoneOtp: otherData.phoneOtp ?? self.phoneOtp
         )
     }
 
@@ -91,6 +99,10 @@ extension StripeAPI.VerificationPageCollectedData {
             self.name = nil
         case .address:
             self.address = nil
+        case .phoneNumber:
+            self.phone = nil
+        case .phoneOtp:
+            self.phoneOtp = nil
         }
     }
 
@@ -135,6 +147,12 @@ extension StripeAPI.VerificationPageCollectedData {
         }
         if self.address != nil {
             ret.insert(.address)
+        }
+        if self.phone != nil {
+            ret.insert(.phoneNumber)
+        }
+        if self.phoneOtp != nil {
+            ret.insert(.phoneOtp)
         }
         return ret
     }