+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/CustomerSheet/CustomerAddPaymentMethodViewController.swift
@@ -11,17 +11,23 @@ import UIKit
 
 protocol CustomerAddPaymentMethodViewControllerDelegate: AnyObject {
     func didUpdate(_ viewController: CustomerAddPaymentMethodViewController)
+    func updateErrorLabel(for: Error?)
 }
 
 @objc(STP_Internal_CustomerAddPaymentMethodViewController)
 class CustomerAddPaymentMethodViewController: UIViewController {
+
+    let paymentMethodTypes: [PaymentSheet.PaymentMethodType]
+
     // MARK: - Read-only Properties
     weak var delegate: CustomerAddPaymentMethodViewControllerDelegate?
-    let paymentMethodTypes: [PaymentSheet.PaymentMethodType] = [.card]
+
     var selectedPaymentMethodType: PaymentSheet.PaymentMethodType {
         return paymentMethodTypesView.selected
     }
     var paymentOption: PaymentOption? {
+        var params = IntentConfirmParams(type: selectedPaymentMethodType)
+        params = paymentMethodFormElement.applyDefaults(params: params)
         if let params = paymentMethodFormElement.updateParams(
             params: IntentConfirmParams(type: selectedPaymentMethodType)
         ) {
@@ -31,8 +37,59 @@ class CustomerAddPaymentMethodViewController: UIViewController {
     }
     // MARK: - Writable Properties
     private let configuration: CustomerSheet.Configuration
+    private lazy var usBankAccountFormElement: USBankAccountPaymentMethodElement? = {
+        // We are keeping usBankAccountInfo in memory to preserve state
+        // if the user switches payment method types
+        let paymentMethodElement = makeElement(for: selectedPaymentMethodType)
+        if let usBankAccountPaymentMethodElement = paymentMethodElement as? USBankAccountPaymentMethodElement {
+            usBankAccountPaymentMethodElement.presentingViewControllerDelegate = self
+        } else {
+            assertionFailure("Wrong type for usBankAccountFormElement")
+        }
+        return paymentMethodElement as? USBankAccountPaymentMethodElement
+    }()
+
+    var overrideActionButtonBehavior: OverrideableBuyButtonBehavior? {
+        if selectedPaymentMethodType == .USBankAccount {
+            if let paymentOption = paymentOption,
+               case .new = paymentOption {
+                return nil // already have PaymentOption
+            } else {
+                return .LinkUSBankAccount
+            }
+        }
+        return nil
+    }
+
+    var overrideCallToAction: ConfirmButton.CallToActionType? {
+        return overrideActionButtonBehavior != nil
+            ? ConfirmButton.CallToActionType.customWithLock(title: String.Localized.continue)
+            : nil
+    }
+    var overrideCallToActionShouldEnable: Bool {
+        guard let overrideBuyButtonBehavior = overrideActionButtonBehavior else {
+            return false
+        }
+        switch overrideBuyButtonBehavior {
+        case .LinkUSBankAccount:
+            return usBankAccountFormElement?.canLinkAccount ?? false
+        }
+    }
+
+    var bottomNoticeAttributedString: NSAttributedString? {
+        if selectedPaymentMethodType == .USBankAccount {
+            if let usBankPaymentMethodElement = paymentMethodFormElement as? USBankAccountPaymentMethodElement {
+                return usBankPaymentMethodElement.mandateString
+            }
+        }
+        return nil
+    }
 
     private lazy var paymentMethodFormElement: PaymentMethodElement = {
+        if selectedPaymentMethodType == .USBankAccount,
+           let usBankAccountFormElement = usBankAccountFormElement {
+            return usBankAccountFormElement
+        }
         return makeElement(for: selectedPaymentMethodType)
     }()
 
@@ -59,10 +116,12 @@ class CustomerAddPaymentMethodViewController: UIViewController {
 
     required init(
         configuration: CustomerSheet.Configuration,
+        paymentMethodTypes: [PaymentSheet.PaymentMethodType],
         delegate: CustomerAddPaymentMethodViewControllerDelegate
     ) {
         self.configuration = configuration
         self.delegate = delegate
+        self.paymentMethodTypes = paymentMethodTypes
         super.init(nibName: nil, bundle: nil)
         self.view.backgroundColor = configuration.appearance.colors.background
     }
@@ -128,7 +187,16 @@ class CustomerAddPaymentMethodViewController: UIViewController {
             }
         }
     }
-
+    private func updateFormElement() {
+        if selectedPaymentMethodType == .USBankAccount,
+           let usBankAccountFormElement = usBankAccountFormElement {
+            paymentMethodFormElement = usBankAccountFormElement
+        } else {
+            paymentMethodFormElement = makeElement(for: selectedPaymentMethodType)
+        }
+        updateUI()
+        sendEventToSubviews(.viewDidAppear, from: view)
+    }
     private func makeElement(for type: PaymentSheet.PaymentMethodType) -> PaymentMethodElement {
         let formElement = PaymentSheetFormFactory(
             configuration: .customerSheet(configuration),
@@ -149,6 +217,69 @@ class CustomerAddPaymentMethodViewController: UIViewController {
     }
 }
 
+extension CustomerAddPaymentMethodViewController {
+    func didTapCallToActionButton(behavior: OverrideableBuyButtonBehavior,
+                                  clientSecret: String,
+                                  from viewController: UIViewController) {
+        switch behavior {
+        case .LinkUSBankAccount:
+            handleCollectBankAccount(from: viewController, clientSecret: clientSecret)
+        }
+    }
+    func handleCollectBankAccount(from viewController: UIViewController, clientSecret: String) {
+        guard
+            let usBankAccountPaymentMethodElement = self.paymentMethodFormElement as? USBankAccountPaymentMethodElement,
+            let name = usBankAccountPaymentMethodElement.name,
+            let email = usBankAccountPaymentMethodElement.email
+        else {
+            assertionFailure()
+            return
+        }
+
+        let params = STPCollectBankAccountParams.collectUSBankAccountParams(
+            with: name,
+            email: email
+        )
+        let client = STPBankAccountCollector()
+        let errorText = STPLocalizedString(
+            "Something went wrong when linking your account.\nPlease try again later.",
+            "Error message when an error case happens when linking your account"
+        )
+        let genericError = PaymentSheetError.unknown(debugDescription: errorText)
+
+        let financialConnectionsCompletion: (FinancialConnectionsSDKResult?, LinkAccountSession?, NSError?) -> Void = {
+            result,
+            _,
+            error in
+            if let error = error {
+                let errorToUse = PaymentSheetError.unknown(debugDescription: error.nonGenericDescription)
+                self.delegate?.updateErrorLabel(for: errorToUse)
+                return
+            }
+            guard let financialConnectionsResult = result else {
+                self.delegate?.updateErrorLabel(for: genericError)
+                return
+            }
+
+            switch financialConnectionsResult {
+            case .cancelled:
+                break
+            case .completed(let linkedBank):
+                usBankAccountPaymentMethodElement.setLinkedBank(linkedBank)
+            case .failed:
+                self.delegate?.updateErrorLabel(for: genericError)
+            }
+        }
+
+        client.collectBankAccountForSetup(
+            clientSecret: clientSecret,
+            returnURL: configuration.returnURL,
+            params: params,
+            from: viewController,
+            financialConnectionsCompletion: financialConnectionsCompletion
+        )
+    }
+}
 extension CustomerAddPaymentMethodViewController: ElementDelegate {
     func continueToNextField(element: Element) {
         delegate?.didUpdate(self)
@@ -162,6 +293,13 @@ extension CustomerAddPaymentMethodViewController: ElementDelegate {
 
 extension CustomerAddPaymentMethodViewController: PaymentMethodTypeCollectionViewDelegate {
     func didUpdateSelection(_ paymentMethodTypeCollectionView: PaymentMethodTypeCollectionView) {
+        updateFormElement()
         delegate?.didUpdate(self)
+    }
+}
+
+extension CustomerAddPaymentMethodViewController: PresentingViewControllerDelegate {
+    func presentViewController(viewController: UIViewController, completion: (() -> Void)?) {
+        self.present(viewController, animated: true, completion: completion)
     }
 }