+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/CustomerSheet/CustomerSheetTests.swift
@@ -19,7 +19,9 @@ class CustomerSheetTests: APIStubbedTestCase {
     func testLoadPaymentMethodInfo_newCustomer() throws {
         let stubbedAPIClient = stubbedAPIClient()
         stubNewCustomerResponse()
+        stubSessions(paymentMethods: "\"card\"")
 
+        let configuration = CustomerSheet.Configuration()
         let customerAdapter = StripeCustomerAdapter(customerEphemeralKeyProvider: {
             .init(customerId: "cus_123", ephemeralKeySecret: "ek_456")
         }, setupIntentClientSecretProvider: {
@@ -27,9 +29,9 @@ class CustomerSheetTests: APIStubbedTestCase {
         }, apiClient: stubbedAPIClient)
 
         let loadPaymentMethodInfo = expectation(description: "loadPaymentMethodInfo completed")
-        let customerSheet = CustomerSheet(configuration: CustomerSheet.Configuration(), customer: customerAdapter)
+        let customerSheet = CustomerSheet(configuration: configuration, customer: customerAdapter)
         customerSheet.loadPaymentMethodInfo { result in
-            guard case .success((let paymentMethods, let selectedPaymentMethod)) = result else {
+            guard case .success((let paymentMethods, let selectedPaymentMethod, _)) = result else {
                 XCTFail()
                 return
             }
@@ -43,7 +45,9 @@ class CustomerSheetTests: APIStubbedTestCase {
     func testLoadPaymentMethodInfo_singleCard() throws {
         let stubbedAPIClient = stubbedAPIClient()
         stubReturningCustomerWithCardResponse()
+        stubSessions(paymentMethods: "\"card\"")
 
+        let configuration = CustomerSheet.Configuration()
         let customerAdapter = StripeCustomerAdapter(customerEphemeralKeyProvider: {
             .init(customerId: "cus_123", ephemeralKeySecret: "ek_456")
         }, setupIntentClientSecretProvider: {
@@ -51,19 +55,79 @@ class CustomerSheetTests: APIStubbedTestCase {
         }, apiClient: stubbedAPIClient)
 
         let loadPaymentMethodInfo = expectation(description: "loadPaymentMethodInfo completed")
-        let customerSheet = CustomerSheet(configuration: CustomerSheet.Configuration(), customer: customerAdapter)
+        let customerSheet = CustomerSheet(configuration: configuration, customer: customerAdapter)
+        customerSheet.loadPaymentMethodInfo { result in
+            guard case .success((let paymentMethods, let selectedPaymentMethod, _)) = result else {
+                XCTFail()
+                return
+            }
+            XCTAssertEqual(paymentMethods.count, 1)
+            XCTAssertEqual(paymentMethods[0].type, .card)
+            XCTAssert(selectedPaymentMethod == nil)
+            loadPaymentMethodInfo.fulfill()
+        }
+        wait(for: [loadPaymentMethodInfo], timeout: 5.0)
+    }
+// TODO: Uncomment when enabling USBankAccount
+/*
+    func testLoadPaymentMethodInfo_singleBankAccount() throws {
+        let stubbedAPIClient = stubbedAPIClient()
+        stubReturningCustomerWithUSBankAccountResponse()
+
+        var configuration = CustomerSheet.Configuration()
+        configuration.paymentMethodTypes = [.USBankAccount]
+        let customerAdapter = StripeCustomerAdapter(customerEphemeralKeyProvider: {
+            .init(customerId: "cus_123", ephemeralKeySecret: "ek_456")
+        }, setupIntentClientSecretProvider: {
+            return "si_789"
+        }, configuration: configuration,
+        apiClient: stubbedAPIClient)
+
+        let loadPaymentMethodInfo = expectation(description: "loadPaymentMethodInfo completed")
+        let customerSheet = CustomerSheet(configuration: configuration, customer: customerAdapter)
         customerSheet.loadPaymentMethodInfo { result in
             guard case .success((let paymentMethods, let selectedPaymentMethod)) = result else {
                 XCTFail()
                 return
             }
             XCTAssertEqual(paymentMethods.count, 1)
+            XCTAssertEqual(paymentMethods[0].type, .USBankAccount)
             XCTAssert(selectedPaymentMethod == nil)
             loadPaymentMethodInfo.fulfill()
         }
         wait(for: [loadPaymentMethodInfo], timeout: 5.0)
     }
 
+    func testLoadPaymentMethodInfo_cardAndBankAccount() throws {
+        let stubbedAPIClient = stubbedAPIClient()
+        stubReturningCustomerWithUSBankAccountResponse()
+        stubReturningCustomerWithCardResponse()
+
+        var configuration = CustomerSheet.Configuration()
+        configuration.paymentMethodTypes = [.card, .USBankAccount]
+        let customerAdapter = StripeCustomerAdapter(customerEphemeralKeyProvider: {
+            .init(customerId: "cus_123", ephemeralKeySecret: "ek_456")
+        }, setupIntentClientSecretProvider: {
+            return "si_789"
+        }, configuration: configuration,
+        apiClient: stubbedAPIClient)
+
+        let loadPaymentMethodInfo = expectation(description: "loadPaymentMethodInfo completed")
+        let customerSheet = CustomerSheet(configuration: configuration, customer: customerAdapter)
+        customerSheet.loadPaymentMethodInfo { result in
+            guard case .success((let paymentMethods, let selectedPaymentMethod)) = result else {
+                XCTFail()
+                return
+            }
+            XCTAssertEqual(paymentMethods.count, 2)
+            XCTAssertEqual(paymentMethods[0].type, .card)
+            XCTAssertEqual(paymentMethods[1].type, .USBankAccount)
+            XCTAssert(selectedPaymentMethod == nil)
+            loadPaymentMethodInfo.fulfill()
+        }
+        wait(for: [loadPaymentMethodInfo], timeout: 5.0)
+    }
+ */
     func testLoadPaymentMethodInfo_CallToPaymentMethodsTimesOut() throws {
         let fastTimeoutIntervalForRequest: TimeInterval = 1
         let timeGreaterThanTimeoutIntervalForRequest: UInt32 = 3
@@ -71,7 +135,9 @@ class CustomerSheetTests: APIStubbedTestCase {
         let stubbedURLSessionConfig = APIStubbedTestCase.stubbedURLSessionConfig()
         stubbedURLSessionConfig.timeoutIntervalForRequest = fastTimeoutIntervalForRequest
         let stubbedAPIClient = stubbedAPIClient(configuration: stubbedURLSessionConfig)
+        stubSessions(paymentMethods: "\"card\"")
 
+        let configuration = CustomerSheet.Configuration()
         let customerAdapter = StripeCustomerAdapter(customerEphemeralKeyProvider: {
             .init(customerId: "cus_123", ephemeralKeySecret: "ek_456")
         }, setupIntentClientSecretProvider: {
@@ -87,7 +153,7 @@ class CustomerSheetTests: APIStubbedTestCase {
         }
 
         let loadPaymentMethodInfo = expectation(description: "loadPaymentMethodInfo completion block called")
-        let customerSheet = CustomerSheet(configuration: CustomerSheet.Configuration(), customer: customerAdapter)
+        let customerSheet = CustomerSheet(configuration: configuration, customer: customerAdapter)
         customerSheet.loadPaymentMethodInfo { result in
             guard case .failure(let error) = result,
                   let nserror = error as NSError?,
@@ -101,19 +167,58 @@ class CustomerSheetTests: APIStubbedTestCase {
         wait(for: [loadPaymentMethodInfo], timeout: 5.0)
     }
 
+    private func stubSessions(paymentMethods: String) {
+        stubSessions(
+            fileMock: .elementsSessionsPaymentMethod_200,
+            responseCallback: { data in
+                return self.updatePaymentMethodDetail(
+                    data: data,
+                    variables: [
+                        "<paymentMethods>": paymentMethods,
+                        "<currency>": "\"usd\"",
+                    ]
+                )
+            }
+        )
+    }
+
+    private func updatePaymentMethodDetail(data: Data, variables: [String: String]) -> Data {
+        var template = String(data: data, encoding: .utf8)!
+        for (templateKey, templateValue) in variables {
+            let translated = template.replacingOccurrences(of: templateKey, with: templateValue)
+            template = translated
+        }
+        return template.data(using: .utf8)!
+    }
+    private func stubSessions(fileMock: FileMock, responseCallback: ((Data) -> Data)? = nil) {
+        stub { urlRequest in
+            return urlRequest.url?.absoluteString.contains("/v1/elements/sessions") ?? false
+        } response: { _ in
+            let mockResponseData = try! fileMock.data()
+            let data = responseCallback?(mockResponseData) ?? mockResponseData
+            return HTTPStubsResponse(data: data, statusCode: 200, headers: nil)
+        }
+    }
+
     private func stubNewCustomerResponse() {
-        stubPaymentMethods(fileMock: .saved_payment_methods_200)
+        stubPaymentMethods(fileMock: .saved_payment_methods_200, pmType: "card")
     }
 
     private func stubReturningCustomerWithCardResponse() {
-        stubPaymentMethods(fileMock: .saved_payment_methods_withCard_200)
+        stubPaymentMethods(fileMock: .saved_payment_methods_withCard_200, pmType: "card")
+    }
+    private func stubReturningCustomerWithUSBankAccountResponse() {
+        stubPaymentMethods(fileMock: .saved_payment_methods_withUSBank_200, pmType: "us_bank_account")
     }
 
     private func stubPaymentMethods(
-        fileMock: FileMock
+        fileMock: FileMock,
+        pmType: String
     ) {
         stub { urlRequest in
-            return urlRequest.url?.absoluteString.contains("/v1/payment_methods") ?? false
+            let isPaymentMethodCall = urlRequest.url?.absoluteString.contains("/v1/payment_methods") ?? false
+            let isPaymentMethodType = urlRequest.url?.absoluteString.contains("type=\(pmType)") ?? false
+            return (isPaymentMethodCall && isPaymentMethodType)
         } response: { _ in
             let mockResponseData = try! fileMock.data()
             return HTTPStubsResponse(data: mockResponseData, statusCode: 200, headers: nil)
@@ -128,4 +233,7 @@ public class ClassForBundle {}
 
     case saved_payment_methods_200 = "MockFiles/saved_payment_methods_200"
     case saved_payment_methods_withCard_200 = "MockFiles/saved_payment_methods_withCard_200"
+    case saved_payment_methods_withUSBank_200 = "MockFiles/saved_payment_methods_withUSBank_200"
+
+    case elementsSessionsPaymentMethod_200 = "MockFiles/elements_sessions_paymentMethod_200"
 }